add server entry & move types to types.ts
