+++ packages/media/src/server.ts
@@ -1,3 +1,5 @@
+import { Breakpoints, BreakpointOptions } from "./types";
+
 /**
  * Creates a very simple and straightforward media query monitor.
  *
@@ -18,6 +20,28 @@ const createMediaQuery = (
   _watchChange: boolean = true
 ): (() => boolean) => {
   return () => false;
+};
+
+/**
+ * Creates a multi-breakpoint monitor to make responsive components easily.
+ * 
+ * @param breakpoints Map of breakpoint names and their widths
+ * @param options Options to customize watch, fallback, responsive mode.
+ * @returns map of currently matching breakpoints.
+ * 
+ * @example
+ * ```ts
+ * const breakpoints = {
+    sm: "640px",
+    lg: "1024px",
+    xl: "1280px",
+  } as const;
+ * const matches = createBreakpoints(breakpoints);
+ * console.log(matches.lg);
+ * ```
+ */
+export const createBreakpoints = (_breakpoints: Breakpoints, options: BreakpointOptions = {}) => {
+  return options.fallbackMatch || {};
 };
 
 export default createMediaQuery;
